split unit and integration tests + upgrade TestFullWorkflow (add scan query use-case)
