+++ .github/workflows/tests.yml
@@ -9,6 +9,32 @@ on:
   pull_request:
   workflow_dispatch:
 jobs:
+  unit:
+    strategy:
+      matrix:
+        go-version: [1.16.x, 1.17.x]
+        os: [ubuntu-latest]
+    env:
+      OS: ${{ matrix.os }}
+      GO: ${{ matrix.go-version }}
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Install Go
+        uses: actions/setup-go@v2
+        with:
+          go-version: ${{ matrix.go-version }}
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Test
+        run: go test -race -coverprofile=unit.txt -covermode=atomic ./...
+      - name: Clear report
+        run: sed -i '/testutil\|trace\|test/d' unit.txt
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v2
+        with:
+          file: ./unit.txt
+          flags: unit,${{ matrix.os }},${{ matrix.go-version }}
+          name: unit
   integration:
     strategy:
       matrix:
@@ -40,7 +66,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Test
-        run: go test -race -coverpkg=./... -coverprofile=integration.txt -covermode=atomic ./...
+        run: go test -race -coverpkg=./... -coverprofile=integration.txt -covermode=atomic ./test/...
       - name: Clear report
         run: sed -i '/testutil\|trace\|test/d' integration.txt
       - name: Upload coverage to Codecov
